# packages/backend/src/generated/server-stubs/routes.ts
import express from 'express';
import { Request, Response } from 'express';

// Auto-generated server stubs based on Swagger specification
// DO NOT MODIFY THIS FILE DIRECTLY

const router = express.Router();


/**
 * Retrieve all articles
 * Path: /api/articles
 * Method: GET
 */
router.get('/articles', (req: Request, res: Response) => {
  // TODO: Implement get_api_articles
  res.status(501).json({ message: 'Not implemented' });
});


/**
 * Create a new article
 * Path: /api/articles
 * Method: POST
 */
router.post('/articles', (req: Request, res: Response) => {
  // TODO: Implement post_api_articles
  res.status(501).json({ message: 'Not implemented' });
});


/**
 * Get article by ID
 * Path: /api/articles/{id}
 * Method: GET
 */
router.get('/articles/{id}', (req: Request, res: Response) => {
  // TODO: Implement get_api_articles__id_
  res.status(501).json({ message: 'Not implemented' });
});


/**
 * Update an existing article
 * Path: /api/articles/{id}
 * Method: PUT
 */
router.put('/articles/{id}', (req: Request, res: Response) => {
  // TODO: Implement put_api_articles__id_
  res.status(501).json({ message: 'Not implemented' });
});


/**
 * Delete an article
 * Path: /api/articles/{id}
 * Method: DELETE
 */
router.delete('/articles/{id}', (req: Request, res: Response) => {
  // TODO: Implement delete_api_articles__id_
  res.status(501).json({ message: 'Not implemented' });
});


export default router;