# packages/backend/src/generated/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Articles-Manager API",
    "version": "1.0.0",
    "description": "API for managing articles"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the article"
          },
          "title": {
            "type": "string",
            "description": "Title of the article"
          },
          "content": {
            "type": "string",
            "description": "Content of the article"
          },
          "author": {
            "type": "string",
            "description": "Author of the article"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with the article"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "author",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the article"
          },
          "content": {
            "type": "string",
            "description": "Content of the article"
          },
          "author": {
            "type": "string",
            "description": "Author of the article"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with the article"
          }
        },
        "required": [
          "title",
          "content",
          "author"
        ]
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the article"
          },
          "content": {
            "type": "string",
            "description": "Content of the article"
          },
          "author": {
            "type": "string",
            "description": "Author of the article"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with the article"
          }
        }
      }
    }
  },
  "paths": {
    "/api/articles": {
      "get": {
        "summary": "Retrieve all articles",
        "tags": [
          "Articles"
        ],
        "description": "Returns a list of all articles in the system",
        "responses": {
          "200": {
            "description": "A list of articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new article",
        "tags": [
          "Articles"
        ],
        "description": "Creates a new article and returns the created article data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Article created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "summary": "Get article by ID",
        "tags": [
          "Articles"
        ],
        "description": "Returns a single article by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The article ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Article details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Article not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing article",
        "tags": [
          "Articles"
        ],
        "description": "Updates an article by ID and returns the updated article",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The article ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Article not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an article",
        "tags": [
          "Articles"
        ],
        "description": "Deletes an article by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The article ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Article deleted successfully (no content)"
          },
          "404": {
            "description": "Article not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Article not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Articles",
      "description": "API endpoints for managing articles"
    }
  ]
}